Keep parsing comparisons
Also, fix a type error on concatenateComparisons.

I was assuming that Token = TokenType, but no. For a moment I
considered using [TokenType] instead of [Token], but that's
rather bad: For literals, I think we might need to whole thing.

We will refactor to do that, though, if it turns out we don't
need it.