Evaluate the minus operator
And to do so, add the concept of a runtime error and attach the
location to expressions. The code involving this is rather
ugly, to be fair. I think that the fact that every possible
value of expression has one `expressionLocation` means we need
to change the data structure. But let's leave that to tomorrow.