Fix he main program
Apparently, the type inference isn't smart enough to work with
type classes instead of concrete types, i.e., I can't convince
them that `(Show a) => Left a` is one whole type. It always
want to concrete `a`. I suspect it's because the types of those
functions is concrete, but let's leave the experimentation
here.