Fix how block behaves
Problem: We were always inserting in the last block.

That means the following: On variable assignment inside a block, if a
variable was declared in another block, it'd create a new one in the
current scope and assign it there.

That would make for loops overly complicated and make while loop
patterns impossible.

To avoid that, I created a "replace" operation for the LoxState:

It will insert the value in the first environment that contains the key,
and in the global scope if none does. This is a horrible thing that will
make lox resemble the worst of Javascript, but I'll fix that later :).