Fix last problem with function calls without return.
When we were parsing expression as arguments, if there were multiple
arguments, it'd confuse comma operator with a comma separator for
expressions. To prevent that, I had to change the grammar to say that
arguments are actually ternary expression (which excludes assignments
and comma expressions). Ideally, I'd change the priority of assignment
and ternary and allow arguments be assignments, but let's leave that for
another day :)