Advance more in the videos
Things I did differently:

- Instead of a double iteration to check cows and bulls, which
is O(n^2), I made use of a dictionary from char to position
under the assumption every letter appears once. That makes the
check O(n) instead.

- For error checking, I created a variant of all the possible
errors and then a visitor for that variant. The method that
checks validity then returns an optional of that variant.